backend/requirements.txt
Django>=4.2
djangorestframework
Pillow
django-cors-headers
